+++ src/pyochain/collections/__init__.py
@@ -5,7 +5,7 @@ from ._deque import Deque
 from ._heapq import Heap, HeapMax, HeapMin
 from ._sorted_dict import SortedDict
 from ._sorted_list import SortedKeyList, SortedList
-from ._sorted_set import SortedSet
+from ._sorted_set import SortedKeySet, SortedSet
 from ._stable_set import StableSet
 
 __all__ = [
@@ -16,6 +16,7 @@ __all__ = [
     "PyoCounter",
     "SortedDict",
     "SortedKeyList",
+    "SortedKeySet",
     "SortedList",
     "SortedSet",
     "StableSet",